POC for modify-config
OK, so I have a proof of concept here for the modify-config functionality referenced in [this issue](https://github.com/McIndi/mast/issues/1).

Still a lot of rough edges, but I believe that it is ready for some field testing. Basically it goes like this:

`--appliances` - Pass in a group of appliances
`--domains` - Pass in a list of domains (will accept `all-domains`)
`--obj-class` - Pass in an object class
`--obj-name` - (optionally) Pass in an object name
`--vector-add` - option is allowed, if specified, the mods will be appended rather than replacing existing values
`--dry-run` - option is allowed, if specified, the `modify-config` requests are not sent to the appliances, rather they are printed to stdout

__Now for the interesting part__:

`--mods` - Define a list of modifications. Multiple options are allowed. They must be in the form of : `--mods "Option=Value"` (if on Linux, in some cases you may need to replace the double-quotes with single-quotes to avoid shell ex